chore(debounce-input): Move to components folder
