+++ www/src/views/starter-library/filtered-starters.js
@@ -21,7 +21,7 @@ import {
   ContentContainer,
 } from "../shared/sidebar"
 import ResetFilters from "../shared/reset-filters"
-import DebounceInput from "./debounce-input"
+import DebounceInput from "../../components/debounce-input"
 
 export default class FilteredStarterLibrary extends Component {
   state = {
@@ -40,7 +40,10 @@ export default class FilteredStarterLibrary extends Component {
   resetFilters = () =>
     this.props.setURLState({ c: null, d: null, v: null, s: `` })
 
-  onChangeUrlWithText = e => this.props.setURLState({ s: e.target.value })
+  onChangeUrlWithText = e => {
+    console.log(e)
+    this.props.setURLState({ s: e.target.value })
+  }
 
   render() {
     const { data, urlState, setURLState } = this.props
@@ -198,7 +201,6 @@ export default class FilteredStarterLibrary extends Component {
                       }`,
                     },
                   }}
-                  type="text"
                   initialValue={urlState.s}
                   onChange={this.onChangeUrlWithText}
                   placeholder="Search starters"
